# api/openapi.yaml
openapi: 3.0.3
info:
  title: qosnat2 API
  description: |
    单机双网卡 QoS（HTB + eBPF + IFB）+ NAT（nftables）+ VPN（WireGuard / ocserv）控制面 REST API。

    **鉴权**：除 `GET /api/v1/health`、`GET /openapi.yaml`、`/api/v1/setup/*` 与 `POST /api/v1/login` 外，需 Session Cookie `qosnat_sess` 或请求头 `X-API-Key`（实验性，见 `/api/v1/api-keys`）。API Key 可设 `role: readonly`，只读密钥禁止写操作（403 `AUTH_READ_ONLY`）。

    **交互文档**：Web UI → 开发 → API (Scalar)，或 `GET /openapi.yaml`。中文说明索引见仓库 `docs/API-ZH.md`。

    **功能域（tags）**：
    - `health` / `setup` / `auth`：健康检查、首次引导、登录与会话
    - `network`：网卡、VLAN/VXLAN、多 WAN、netplan 预览与应用
    - `nat`：SNAT/DNAT、防火墙规则与别名对象组
    - `shaper`：QoS 模板、租户、HTB 叶子队列
    - `stats` / `ebpf` / `system`：仪表盘、BPF、sysctl/HTTPS/审计
    - `vpn`：WireGuard、ocserv（OpenConnect）含虚拟主机与独立用户
    - `diagnostics`：tcpdump 抓包、conntrack 列表

    **ocserv 常用流程**：
    1. `GET /api/v1/vpn/ocserv` 读取配置与 `install_job` / `status`
    2. `PUT` 保存 state（密码/RADIUS 密钥留空表示不修改）
    3. `POST /api/v1/vpn/ocserv/apply` 写入 `ocserv.conf`；运行中仅 `reload`，冷启动用 `start`，关闭用 `stop`+`disable`（非 reload 生效项见 GET `restart_pending_reasons`）
    4. `POST /api/v1/vpn/ocserv/service`：`start` / `stop` / `restart`（需 root）
    5. 虚拟主机：`/api/v1/vpn/ocserv/vhosts`；独立密码文件用户：`/api/v1/vpn/ocserv/vhosts/users?domain=`
    6. 安装：`POST .../install`（202，需 root；源码编译）→ 轮询 `GET .../install/status`
    7. 卸载：`POST .../uninstall`（需 root + Web 管理员密码 `admin_password`）

...
  version: 1.0.0
  contact:
    name: qosnat2
servers:
  - url: /
    description: 当前 qosnatd 实例
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: qosnat_sess
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - sessionAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/setup/complete:
    post:
      tags: [setup]
      summary: 完成初始设置（管理员、LAN/WAN、可选 SNAT IP）
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [admin_user, admin_pass, dev_lan, dev_wan]
              properties:
                admin_user: { type: string }
                admin_pass: { type: string }
                dev_lan: { type: string }
                dev_wan: { type: string }
                policy_routes: { type: array, items: { type: string } }
                shared_ips: { type: array, items: { type: string } }
                hostname: { type: string }
                enable_dhcp: { type: boolean }
                apply_dataplane: { type: boolean }
      responses:
        '200':
          description: OK（Set-Cookie 登录）
  /api/v1/login:
    post:
      tags: [auth]
      summary: 登录（Set-Cookie qosnat_sess）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user, pass]
              properties:
                user: { type: string }
                pass: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/api-keys:
    get:
      tags: [auth]
      summary: 列出 API Key（实验性，不含完整 secret）
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string }
                    name: { type: string }
                    role: { type: string, enum: [admin, readonly, firewall] }
                    key_prefix: { type: string }
                    created_at: { type: string }
    post:
      tags: [auth]
      summary: 创建 API Key（仅创建响应返回完整 key）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                role:
                  type: string
                  enum: [admin, readonly, firewall]
                  default: admin
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  name: { type: string }
                  role: { type: string }
                  key: { type: string }
                  created_at: { type: string }
        '403':
          description: 只读 API Key 禁止写操作（AUTH_READ_ONLY）
    delete:
      tags: [auth]
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
  /api/v1/dhcp:
    get:
      tags: [network]
      summary: DHCP 配置、dnsmasq 状态与网卡列表
      responses:
        '200':
          description: config + status + interfaces + dev_lan/dev_wan
    put:
      tags: [network]
      summary: 保存 DHCP 配置（不立即应用）
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DHCPState'
      responses:
        '200':
          description: OK
  /api/v1/dhcp/dnsmasq/install-chnroutes:
    post:
      tags: [network]
      summary: 编译安装 chnroutes 补丁版 dnsmasq（需 root，后台任务）
      responses:
        '202':
          description: accepted + job
        '200':
          description: 已支持 chnroutes，无需安装
...
  /api/v1/dhcp/chnroutes/update:
    post:
      tags: [network]
      summary: 下载/更新 chnroutes CIDR 列表
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                url: { type: string, description: "下载源 URL，默认 hk59775634/chnroutes（jsDelivr，失败时回退 GitHub Raw）" }
                path: { type: string, description: 写入路径，默认 state 中 chnroutes_file }
      responses:
        '200':
          description: ok + entries + chnroutes info
...
  /api/v1/routes:
    get:
      tags: [network]
      summary: 托管路由 + 内核 main 表快照
      responses:
        '200':
          description: managed + live + dev_lan/dev_wan
    post:
      tags: [network]
      summary: 添加托管静态路由并 ip route replace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteEntry'
      responses:
        '200':
          description: 创建的条目
...
  /api/v1/frr:
    get:
      tags: [network]
      summary: FRR 状态、路由后端与配置摘要
      responses:
        '200':
          description: status, route_backend, boot_on_startup, extra_config
    put:
      tags: [network]
      summary: 保存路由后端与开机启动设置
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                route_backend: { type: string, enum: [kernel, frr] }
                boot_on_startup: { type: boolean }
                extra_config: { type: string }
      responses:
        '200':
          description: ok
  /api/v1/frr/install:
    post:
      tags: [network]
      summary: 后台 apt 安装 FRR
      responses:
        '202':
          description: install started
        '200':
          description: already installed
...
  /api/v1/frr/service:
    post:
      tags: [network]
      summary: FRR systemd 操作 start|stop|restart
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [start, stop, restart] }
      responses:
        '200':
          description: ok
  /api/v1/frr/config:
    get:
      tags: [network]
      summary: 读取 FRR 原始配置文件
      parameters:
        - name: which
          in: query
          schema: { type: string, enum: [frr.conf, extra, managed, daemons, include] }
      responses:
        '200':
          description: path + content
    put:
      tags: [network]
      summary: 写入 FRR 原始配置文件
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [which, content]
              properties:
                which: { type: string }
                content: { type: string }
      responses:
        '200':
          description: ok
...
  /api/v1/nat/policy-routes:
    get:
      tags: [nat]
      summary: SNAT 策略网段列表
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
    post:
      tags: [nat]
      summary: 添加策略网段
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [cidr]
              properties:
                cidr: { type: string, example: 10.0.0.0/8 }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
    delete:
      tags: [nat]
      summary: 删除策略网段
      parameters:
        - name: cidr
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
  /api/v1/nat/shared-ips:
    get:
      tags: [nat]
      summary: 共享 SNAT 公网 IP 池（未配置时 ips 为 WAN 口 IPv4）
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  ips: { type: array, items: { type: string } }
                  configured: { type: array, items: { type: string } }
                  auto_from_wan: { type: boolean }
                  wan_device: { type: string }
    post:
      tags: [nat]
      summary: 添加共享 IP
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [ip]
              properties:
                ip: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
    delete:
      tags: [nat]
      parameters:
        - name: ip
          in: query
          required: true
          schema: { type: string }
...
  /api/v1/nat/nptv6:
    get:
      tags: [nat]
      summary: NPTv6 配置
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  nptv6_enabled: { type: boolean }
                  nptv6_rules:
                    type: array
                    items: { $ref: '#/components/schemas/Nptv6Rule' }
    put:
      tags: [nat]
      summary: 更新 NPTv6 并 apply
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nptv6_enabled: { type: boolean }
                nptv6_rules:
                  type: array
                  items: { $ref: '#/components/schemas/Nptv6Rule' }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
  /api/v1/nat/nat64:
    get:
      tags: [nat]
      summary: NAT64 + DNS64 配置与运行状态
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    put:
      tags: [nat]
      summary: 更新 NAT64/DNS64 并 apply（Jool + Unbound + dnsmasq）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nat64_enabled: { type: boolean }
                nat64_prefix: { type: string }
                nat64_pool4: { type: string }
                dns64: { $ref: '#/components/schemas/DNS64Config' }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
  /api/v1/nat/dns64:
    get:
      tags: [nat]
      summary: DNS64 子配置
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DNS64Config'
    put:
      tags: [nat]
      summary: 仅更新 DNS64 并 apply
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DNS64Config'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
  /api/v1/nat/wan-forwards:
    get:
      tags: [nat]
      summary: 端口转发列表 + 网卡/默认项
      responses:
        '200':
          description: forwards, interfaces, defaults, dev_lan/dev_wan
    post:
      tags: [nat]
      summary: 添加 DNAT 规则并 reload nft
      description: |
        同步生成 prerouting DNAT、forward 链 `auto-fwd-{id}-{proto}` 放行规则与 LAN hairpin NAT。
        删除转发会一并移除上述自动防火墙规则。
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WanPortForward'
      responses:
        '200':
          description: 创建的规则
        '400':
          description: 校验失败或 nft 预检失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [nat]
      summary: 按 id 删除规则
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
  /api/v1/shaper/enabled:
    get:
      tags: [shaper]
      summary: QoS 总开关状态
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled: { type: boolean }
    put:
      tags: [shaper]
      summary: 启用/关闭 QoS（关闭时清除运行中 TC/eBPF/ifb 整形，保留已保存策略）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
                apply: { type: boolean, description: 立即应用到数据面 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  enabled: { type: boolean }
...
  /api/v1/system/certificates/renew:
    post:
      tags: [system]
      summary: ACME 续签证书库条目（需 root）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string }
      responses:
        '200':
          description: 已续签
...